split tests for deprecated api (#5071)
* imports

* imports

* flake8
